refactor: Extract useDeployGasLimit hook
